Refactor: extension이므로 extensions 폴더로 파일 이동 시킴
